+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   * Option to display stats for each swap device separately
   * Dynamically load / unload swap devices as they are added / disabled
   * Notify user when they try to send signal to processes owned by different UID
+  * Run new task option
+  * Open terminal option
 
 # 1.0.3
   * Performance improvements (app uses less RAM and CPU now)